自定义数组测试
