--- src/test/java/org/lxg/data/DataTest.java
@@ -1,4 +1,0 @@
-package org.lxg.data;
-
-public class DataTest {
-}
