--- src/test/java/org/lxg/data/array/SelfArray.java
@@ -1,32 +1,0 @@
-package org.lxg.data.array;
-
-public class SelfArray {
-    private int[] data;
-    private int size;
-
-    // 构造函数，传入数组的容量capacity构造Array
-    public SelfArray(int capacity){
-        data = new int[capacity];
-        size = 0;
-    }
-
-    // 无参数的构造函数，默认数组的容量capacity=10
-    public SelfArray(){
-        this(10);
-    }
-
-    // 获取数组的容量
-    public int getCapacity(){
-        return data.length;
-    }
-
-    // 获取数组中的元素个数
-    public int getSize(){
-        return size;
-    }
-
-    // 返回数组是否为空
-    public boolean isEmpty(){
-        return size == 0;
-    }
-}
